==> 04. Evaluating models.ipynb <==
import pandas as pd
import glob

import numpy as np
results = pd.concat([pd.read_csv(f) for f in glob.glob('data/result*')], ignore_index=True)
results = results.groupby(["experiment", "hierarchy"]).agg(
    in_top_one=('in_top_one', "sum"), 
    in_top_three=('in_top_three', "sum"), 
    in_top_ten=('in_top_ten', "sum"), 
    n = ('in_top_ten', "count")
).reset_index()

results.in_top_one = results.in_top_one / results.n
results.in_top_three = results.in_top_three / results.n
results.in_top_ten = results.in_top_ten / results.n
results.to_csv("data/evaluation.csv")
results = results.sort_values(["hierarchy", "experiment"])
results[results.hierarchy == "baseterm"]
